Added the following patch from the dev@openidenabled.com mailing list: http://lists.openidenabled.com/pipermail/dev/attachments/20090109/7e344691/attachment-0001.bin
Original Message:
ketmar at ketmar.no-ip.org ketmar at ketmar.no-ip.org
Fri Jan 9 07:51:35 PST 2009
darcs patch: fix for incomplete URIs in "location" http field (for ...

	"* fix for incomplete URIs in "location" http field (for technorati and maybe others)"

This patch was in the form of a Darcs patch, not a normal patch.  So solve this, I applied it to the Darcs repository found on openidenabled, then created a new diff file between the original Darcs repo and the new one (with the patch applied) so that I could apply it to this git repo.

Hunks were applied successfully.